+++ src/assets/img/requiredThings/Flashlight.png
