+++ src/assets/img/requiredThings/Gloves.png
