+++ src/assets/img/requiredThings/Scarf.png
